Fix last commit
Issue: My first commit was made after I had already
edited the exercises original code. After reverting back
to the first commit, there were still traces of my initial commit.

Solution: Start creating an empty 'init' commit in the
case a revert to a previous commit or the initial commit
is needed in the future. (As it most likely will be)